+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rrweb/types
 
+## 2.0.0-alpha.10
+
+### Patch Changes
+
+- [#1268](https://github.com/rrweb-io/rrweb/pull/1268) [`d872d28`](https://github.com/rrweb-io/rrweb/commit/d872d2809e3ec8d6ff5d3d5f43bc81aff70e7548) Thanks [@eoghanmurray](https://github.com/eoghanmurray)! - Compact style mutation fixes and improvements
+
+  - fixes when style updates contain a 'var()' on a shorthand property #1246
+  - further ensures that style mutations are compact by reverting to string method if it is shorter
+
+- Updated dependencies [[`c6600e7`](https://github.com/rrweb-io/rrweb/commit/c6600e742b8ec0b6295816bb5de9edcd624d975e)]:
+  - rrweb-snapshot@2.0.0-alpha.10
+
 ## 2.0.0-alpha.9
 
 ### Patch Changes